move training program create schema to schema file
